+++ src/test/java/org/cicirello/math/rand/RandomIndexerTests.java
@@ -36,210 +36,6 @@ public class RandomIndexerTests extends SharedTestRandom {
 	private static double EPSILON = 1e-10;
 		
 	@Test
-	public void testArrayMask() {
-		SplittableRandom r1 = new SplittableRandom(42);
-		for (int n = 1; n <= 5; n++) {
-			for (int k = 0; k <= n; k++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, k);
-				assertEquals( n, mask.length);
-				int count = 0;
-				for (int i = 0; i < mask.length; i++) {
-					if (mask[i]) count++;
-				}
-				assertEquals( k, count);
-				
-				mask = RandomIndexer.arrayMask(n, k, r1);
-				assertEquals( n, mask.length);
-				count = 0;
-				for (int i = 0; i < mask.length; i++) {
-					if (mask[i]) count++;
-				}
-				assertEquals( k, count);
-				
-				mask = RandomIndexer.arrayMask(n);
-				assertEquals( n, mask.length);
-				
-				mask = RandomIndexer.arrayMask(n, r1);
-				assertEquals( n, mask.length);
-			}
-		}
-		
-		final int TRIALS = 120000;
-		for (int n = 1; n <= 5; n++) {
-			final int MAX = TRIALS / n;
-			int[] buckets = {0, 0};
-			for (int t = 0; t < MAX; t++) {
-				boolean[] mask = RandomIndexer.arrayMask(n);
-				for (int i = 0; i < n; i++) {
-					if (mask[i]) buckets[1]++;
-					else buckets[0]++;
-				}
-			}
-			double chi = chiSquare(buckets);
-			assertTrue( chi <= 3.841, "Using ThreadLocalRandom, chi square goodness of fit test: " + chi);
-		}
-		for (int n = 1; n <= 5; n++) {
-			final int MAX = TRIALS / n;
-			int[] buckets = {0, 0};
-			for (int t = 0; t < MAX; t++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, r1);
-				for (int i = 0; i < n; i++) {
-					if (mask[i]) buckets[1]++;
-					else buckets[0]++;
-				}
-			}
-			double chi = chiSquare(buckets);
-			assertTrue( chi <= 3.841, "Using SplittableRandom, chi square goodness of fit test: " + chi);
-		}
-	}
-	
-	@Test
-	public void testArrayMaskP05() {
-		double p = 0.5;
-		SplittableRandom r1 = new SplittableRandom(42);
-		for (int n = 1; n <= 5; n++) {
-			for (int k = 0; k <= n; k++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p);
-				assertEquals( n, mask.length);
-				
-				mask = RandomIndexer.arrayMask(n, p, r1);
-				assertEquals( n, mask.length);
-			}
-		}
-		
-		final int TRIALS = 120000;
-		for (int n = 1; n <= 5; n++) {
-			final int MAX = TRIALS / n;
-			int[] buckets = {0, 0};
-			for (int t = 0; t < MAX; t++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p);
-				for (int i = 0; i < n; i++) {
-					if (mask[i]) buckets[1]++;
-					else buckets[0]++;
-				}
-			}
-			double chi = chiSquare(buckets);
-			assertTrue( chi <= 3.841, "Using ThreadLocalRandom, chi square goodness of fit test: " + chi);
-		}
-		for (int n = 1; n <= 5; n++) {
-			final int MAX = TRIALS / n;
-			int[] buckets = {0, 0};
-			for (int t = 0; t < MAX; t++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p, r1);
-				for (int i = 0; i < n; i++) {
-					if (mask[i]) buckets[1]++;
-					else buckets[0]++;
-				}
-			}
-			double chi = chiSquare(buckets);
-			assertTrue( chi <= 3.841, "Using SplittableRandom, chi square goodness of fit test: " + chi);
-		}
-	}
-	
-	@Test
-	public void testArrayMaskP075() {
-		double p = 0.75;
-		double[] pa = {0.25, 0.75};
-		SplittableRandom r1 = new SplittableRandom(42);
-		for (int n = 1; n <= 5; n++) {
-			for (int k = 0; k <= n; k++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p);
-				assertEquals( n, mask.length);
-				
-				mask = RandomIndexer.arrayMask(n, p, r1);
-				assertEquals( n, mask.length);
-			}
-		}
-		
-		final int TRIALS = 120000;
-		for (int n = 1; n <= 5; n++) {
-			final int MAX = TRIALS / n;
-			int[] buckets = {0, 0};
-			for (int t = 0; t < MAX; t++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p);
-				for (int i = 0; i < n; i++) {
-					if (mask[i]) buckets[1]++;
-					else buckets[0]++;
-				}
-			}
-			double chi = chiSquare(buckets, pa);
-			assertTrue( chi <= 3.841, "Using ThreadLocalRandom, chi square goodness of fit test: " + chi);
-		}
-		for (int n = 1; n <= 5; n++) {
-			final int MAX = TRIALS / n;
-			int[] buckets = {0, 0};
-			for (int t = 0; t < MAX; t++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p, r1);
-				for (int i = 0; i < n; i++) {
-					if (mask[i]) buckets[1]++;
-					else buckets[0]++;
-				}
-			}
-			double chi = chiSquare(buckets, pa);
-			assertTrue( chi <= 3.841, "Using SplittableRandom, chi square goodness of fit test: " + chi);
-		}
-	}
-	
-	@Test
-	public void testArrayMaskP025() {
-		double p = 0.25;
-		double[] pa = {0.75, 0.25};
-		SplittableRandom r1 = new SplittableRandom(42);
-		for (int n = 1; n <= 5; n++) {
-			for (int k = 0; k <= n; k++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p);
-				assertEquals( n, mask.length);
-				
-				mask = RandomIndexer.arrayMask(n, p, r1);
-				assertEquals( n, mask.length);
-			}
-		}
-		
-		final int TRIALS = 120000;
-		for (int n = 1; n <= 5; n++) {
-			final int MAX = TRIALS / n;
-			int[] buckets = {0, 0};
-			for (int t = 0; t < MAX; t++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p);
-				for (int i = 0; i < n; i++) {
-					if (mask[i]) buckets[1]++;
-					else buckets[0]++;
-				}
-			}
-			double chi = chiSquare(buckets, pa);
-			assertTrue( chi <= 3.841, "Using ThreadLocalRandom, chi square goodness of fit test: " + chi);
-		}
-		for (int n = 1; n <= 5; n++) {
-			final int MAX = TRIALS / n;
-			int[] buckets = {0, 0};
-			for (int t = 0; t < MAX; t++) {
-				boolean[] mask = RandomIndexer.arrayMask(n, p, r1);
-				for (int i = 0; i < n; i++) {
-					if (mask[i]) buckets[1]++;
-					else buckets[0]++;
-				}
-			}
-			double chi = chiSquare(buckets, pa);
-			assertTrue( chi <= 3.841, "Using SplittableRandom, chi square goodness of fit test: " + chi);
-		}
-	}
-	
-	@Test
-	public void testArrayMaskBoundaryCases() {
-		SplittableRandom r1 = new SplittableRandom(42);
-		boolean[] mask = RandomIndexer.arrayMask(10, 1.0, r1);
-		assertEquals(10, mask.length);
-		for (int i = 0; i < mask.length; i++) {
-			assertTrue(mask[i]);
-		}
-		mask = RandomIndexer.arrayMask(10, 0.0, r1);
-		assertEquals(10, mask.length);
-		for (int i = 0; i < mask.length; i++) {
-			assertFalse(mask[i]);
-		}
-	}
-	
-	@Test
 	public void testRandInt_ThreadLocalRandom() {
 		final int REPS_PER_BUCKET = 100;
 		double[] limit95 = {
